All updates InputSource makes to the layer's dimensions are passed through NSIntegralRect to stop crazy floating point coordinates from being set.
When searching for an InputSource to select on mouseover/click, if the topmost input has a parent and their frames are equal, fall through to the parent(s). This change means that when you attach an input to another input and then auto-fit it, click-drag, resize and right click all use the PARENT layer.